Move check object method - Keep the organization - Move to objects module
